api_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
compressed_rtf==1.0.6
cryptography==42.0.5
distro==1.9.0
docx==0.2.4
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
IMAPClient==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
olefile==0.47
openai==1.13.3
pdfminer.six==20191110
pillow==10.2.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
python-docx==1.1.0
python-dotenv==1.0.1
python-pptx==0.6.23
regex==2023.12.25
requests==2.31.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
textract==1.6.5
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
xlrd==1.2.0
XlsxWriter==3.2.0
